Refactor button css and split button from link button
