+++ src/components/CloseButton.tsx
@@ -1,0 +1,5 @@
+import Link from 'next/link';
+
+export default function CloseButton() {
+  return <Link href="/" className="close-button" aria-label="Close" />;
+}
