--- src/components/CloseButton/index.tsx
@@ -1,6 +1,0 @@
-import Link from 'next/link';
-import styles from './closebutton.module.scss';
-
-export default function CloseButton() {
-  return <Link href="/" className={styles.button} aria-label="Close" />;
-}
