+++ src/styles/globals.scss
@@ -95,4 +95,6 @@ figure {
   }
 }
 
+@import './modules/button.scss';
+@import './modules/close-button.scss';
 @import './modules/event-list.scss';
