+++ src/styles/modules/close-button.scss
@@ -1,4 +1,4 @@
-.button {
+.close-button {
   position: absolute;
   top: 1.5rem;
   right: 1.5rem;
